fix: address PR review feedback for reserved-failed sandbox handling
Move IsReservedFailedSandbox to pkg/utils, filter reserved-failed sandboxes
in ListSandboxes, and harden failed-sandbox cleanup logging and deletion.
