+++ pkg/servers/e2b/list.go
@@ -33,6 +33,7 @@ import (
 	"github.com/openkruise/agents/pkg/sandbox-manager/infra"
 	"github.com/openkruise/agents/pkg/servers/e2b/models"
 	"github.com/openkruise/agents/pkg/servers/web"
+	"github.com/openkruise/agents/pkg/utils"
 	"github.com/openkruise/agents/pkg/utils/pagination"
 )
 
@@ -183,7 +184,7 @@ func getListFilter(request ListSandboxesRequest) func(sbx infra.Sandbox) bool {
 		request.States = []string{agentsv1alpha1.SandboxStateRunning, agentsv1alpha1.SandboxStatePaused}
 	}
 	return func(sbx infra.Sandbox) bool {
-		if isReservedFailedSandbox(sbx) {
+		if utils.IsReservedFailedSandbox(sbx.GetLabels()) {
 			return false
 		}
 		if len(request.States) > 0 {
